requirements.txt
numpy==1.18.4
Pillow==8.2.0
tqdm==4.46.0
opencv-python==4.2.0.34
addict==2.2.1
termcolor==1.1.0
pyclipper==1.1.0.post3
shapely==1.7.0
torch>=1.7.0
torchvision>=0.8.0
python-Levenshtein>=0.12.0
lmdb>=0.98
imgaug>=0.4.0
line-profiler==3.2.6
memory-profiler==0.58.0
